server/requirements.txt
Flask==3.0.3
flask-cors
gunicorn==22.0.0
firebase_admin==6.4.0
celery==5.4.0
redis==5.2.1
httpx>=0.27.2
svix
google-cloud-firestore
